Wrote new gpgmailer.py
